add torch-tilt-series to monorepo (renamed from torch-tomogram)
- Add torch-tomogram via git subtree from https://github.com/teamtomo/torch-tomogram
- Rename package to torch-tilt-series throughout:
  - Directory: packages/primitives/torch-tilt-series/
  - Module: src/torch_tilt_series/
  - Package name in pyproject.toml
  - All imports updated from torch_tomogram to torch_tilt_series
  - Coverage source configuration
- Update pyproject.toml for monorepo compatibility (tag-pattern, URLs, Python >=3.12)
- Clean up monorepo-level config files (.github, .gitignore, .pre-commit-config.yaml)
- Update LICENSE to TeamTomo copyright
- Add to root dependencies and workspace sources as torch-tilt-series
- Tests passing: 5 passed, 1 skipped in 1.31s
