--- packages/primitives/torch-tilt-series/.github/CODEOWNERS
@@ -1,2 +1,0 @@
-# HARDEN .github directory (from the root of the repo) to only be altered after review by maintainers
-/.github/ @alisterburt @McHaillet
