--- packages/primitives/torch-tilt-series/.github/ISSUE_TEMPLATE.md
@@ -1,15 +1,0 @@
-* torch-tomogram version:
-* Python version:
-* Operating System:
-
-### Description
-
-Describe what you were trying to get done.
-Tell us what happened, what went wrong, and what you expected to happen.
-
-### What I Did
-
-```
-Paste the command(s) you ran and the output.
-If there was a crash, please include the traceback here.
-```
