+++ packages/primitives/torch-tilt-series/LICENSE
@@ -1,6 +1,7 @@
 BSD 3-Clause License
 
-Copyright (c) 2023, Talley Lambert
+Copyright (c) 2020, TeamTomo
+All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
