+++ packages/primitives/torch-tilt-series/src/torch_tilt_series/__init__.py
@@ -3,12 +3,12 @@
 from importlib.metadata import PackageNotFoundError, version
 
 try:
-    __version__ = version("torch-tomogram")
+    __version__ = version("torch-tilt-series")
 except PackageNotFoundError:
     __version__ = "uninstalled"
 __author__ = "Marten Chaillet"
 __email__ = "martenchaillet@gmail.com"
 
-from torch_tomogram.tomogram import Tomogram
+from torch_tilt_series.tomogram import Tomogram
 
 __all__ = ["Tomogram"]
